Task5-flask/water_segmentation_app/requirements.txt
# Core dependencies
flask==2.0.1
werkzeug==2.0.2
numpy==1.22.3
matplotlib==3.5.1
pillow==9.0.1
tifffile==2022.2.9
tensorflow
protobuf

# Other utilities tat might be helpful
opencv-python-headless==4.5.5.64
scikit-learn==1.0.2